+++ packages/client/src/assets/styles/index.scss
@@ -13,12 +13,12 @@ div[id='root'] {
   display: flex;
   flex-direction: column;
   border: 0;
+  overflow: hidden;
   font-family: Quicksand, sans-serif;
   font-size: 16px;
   font-weight: 400;
   font-style: normal;
   scroll-behavior: smooth;
-  overflow: hidden;
 }
 
 body {
