+++ packages/client/src/components/AvatarImage/styles.module.scss
@@ -8,10 +8,10 @@
   justify-content: center;
   border-radius: 50%;
   background: $color-palette-white;
-  min-width: 70px;
-  min-height: 70px;
   width: 70px;
+  min-width: 70px;
   height: 70px;
+  min-height: 70px;
 }
 
 .imageContainerSmall {
@@ -20,10 +20,10 @@
   justify-content: center;
   border-radius: 50%;
   background: $color-palette-white;
-  min-width: 30px;
-  min-height: 30px;
   width: 30px;
+  min-width: 30px;
   height: 30px;
+  min-height: 30px;
 }
 
 .imageDefault {
