+++ packages/client/src/components/CommentItem/styles.module.scss
@@ -15,13 +15,15 @@
 }
 
 .username {
-  color: $color-palette-blue;
   @include typography-m;
+
+  color: $color-palette-blue;
 }
 
 .text {
-  color: $color-palette-blue;
   @include typography-m;
+
+  color: $color-palette-blue;
 }
 
 .contentContainer {
@@ -33,8 +35,8 @@
 
 .textContainer {
   display: flex;
-  width: 100%;
-  min-height: 40px;
   border-left: 2px solid $color-palette-white;
   padding: 0 0 0 5px;
+  width: 100%;
+  min-height: 40px;
 }
