+++ packages/client/src/components/TopicCreationForm/styles.module.scss
@@ -1,8 +1,55 @@
+@import '../../assets/styles/mixins';
+@import '../../assets/styles/variables';
 
-
-.messageInput {
-  min-height: 100px;
+.group {
   display: flex;
-  align-items: flex-start;
-  justify-content: flex-start;
+  position: relative;
+  flex-direction: column;
+  margin-bottom: 20px;
+}
+
+.label {
+  margin-bottom: 15px;
+  line-height: 9px;
+  font-weight: 700;
+}
+
+.textArea {
+  @include typography-xs;
+
+  position: relative;
+  border: none;
+  background: none;
+  padding-bottom: 5px;
+  padding-left: 10px;
+  width: 100%;
+  max-width: 100%;
+  min-height: 135px;
+  max-height: 50vh;
+  color: $color-palette-white;
+  font-family: 'Press Start 2P', sans-serif;
+  font-weight: 300;
+}
+
+.textArea::placeholder {
+  background: none;
+  font-family: Quicksand, sans-serif;
+  font-weight: 300;
+}
+
+.textArea::after {
+  position: absolute;
+  bottom: 0;
+  left: 0;
+  box-shadow: 0 2px 4px 0 $color-palette-yellow;
+  background: $color-palette-white;
+  width: 100%;
+  height: 1px;
+  content: '';
+}
+
+textarea:-webkit-autofill {
+  -webkit-text-fill-color: $color-palette-white;
+  box-shadow: 0 0 0 40px black inset;
+  font-family: 'Press Start 2P', sans-serif !important;
 }
