+++ packages/client/src/components/TopicItem/styles.module.scss
@@ -7,44 +7,37 @@
 
 .imageContainer {
   display: flex;
-  width: 100%;
   border-bottom: 6px solid $color-palette-blue;
   padding: 7px 10px;
-}
-
-.image {
-  width: 35.07px;
-  height: 37px;
+  width: 100%;
 }
 
 .content {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
   width: 100%;
   height: 70px;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
 }
 
 .text {
   @include typography-l;
+
   font-family: $font-Press-start-2p;
 }
 
 .linkImage {
-  padding: 5px 0 0 0;
+  padding: 5px 0 0;
 }
 
 .topicGroup {
   display: flex;
-  align-items: center;
   gap: 104px;
+  align-items: center;
   margin: 0 10px 0 0;
 }
 
 .counter {
   color: $color-palette-green;
   font-size: 36px;
-  font-style: normal;
-  font-weight: 400;
-  line-height: normal;
 }
