+++ packages/client/src/components/TopicMessage/styles.module.scss
@@ -3,15 +3,15 @@
 
 .container {
   display: flex;
-  width: 100%;
   flex-direction: column;
   gap: 10px;
+  width: 100%;
 }
 
 .userContainer {
   display: flex;
-  width: 100%;
   gap: 21px;
+  width: 100%;
 }
 
 .loginContainer {
@@ -21,29 +21,31 @@
 }
 
 .username{
-  color: $color-palette-blue;
   @include typography-l;
+
   display: flex;
   padding: 10px;
+  color: $color-palette-blue;
 }
 
 
 .messageContainer {
+  display: flex;
+  flex-direction: column;
+  gap: 30px;
   border-left: 2px solid #FFF;
   padding: 0 0 0 15px;
   width: 100%;
   height: 100%;
-  display: flex;
-  flex-direction: column;
-  gap: 30px;
 }
 
 .message {
-  min-height: 70px;
+  @include typography-m;
+
   display: flex;
   align-items: flex-start;
   width: 100%;
-  @include typography-m;
+  min-height: 70px;
   color: $color-palette-blue;
 }
 
@@ -60,9 +62,10 @@
 }
 
 .commentsLength {
-  color: $color-palette-blue;
   @include typography-s;
+
   display: flex;
   gap: 5px;
   cursor: pointer;
+  color: $color-palette-blue;
 }
