+++ packages/client/src/pages/Topic/styles.module.scss
@@ -3,29 +3,30 @@
 
 .container {
   display: flex;
-  width: 100%;
   flex-direction: column;
+  width: 100%;
   height: 100%;
   overflow: hidden;
 }
 
 .textContainer {
-  width: 100%;
   display: flex;
   justify-content: space-between;
-  margin: 0 0 42px 0;
+  margin: 0 0 42px;
+  width: 100%;
 }
 
 .headingContainer {
   display: flex;
-  align-items: center;
   gap: 28px;
+  align-items: center;
 }
 
 .heading {
-  font-family: $font-Press-start-2p;
   @include typography-xl;
+
   color: $color-palette-blue;
+  font-family: $font-Press-start-2p;
 }
 
 .image {
@@ -34,45 +35,41 @@
 }
 
 .backButton {
-  font-size: 24px;
-  font-style: normal;
-  font-weight: 400;
-  line-height: normal;
-  background: none;
-  border: none;
-  color: $color-palette-green;
   display: flex;
   align-items: flex-end;
+  border: none;
+  background: none;
   padding: 0;
+  color: $color-palette-green;
+  font-size: 24px;
 }
 
 .messages {
-  border-radius: 20px 10px 10px 20px;
-  background: $color-palette-green;
   display: flex;
   flex-direction: column;
-  width: 100%;
+  border-radius: 20px 10px 10px 20px;
+  background: $color-palette-green;
+  padding: 0 0 29px;
   height: 72vh;
-  padding: 0 0 29px 0;
 }
 
 .messageList {
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-  padding: 19px 66px 25px 19px;
-  position: relative;
   @include scrollbars;
-  overflow-y: scroll;
-  overflow-x: hidden;
+
+  display: flex;
+  position: relative;
+  flex-direction: column;
   gap: 25px;
+  padding: 19px 66px 25px 19px;
+  height: 100%;
+  overflow-x: hidden;
+  overflow-y: scroll;
 }
 
 .formContainer {
   display: flex;
-  min-height: 70px;
   align-items: flex-end;
-  width: 100%;
-  padding: 19px 19px 0 19px;
-  margin: 0 0 31px 0;
+  margin: 0 0 31px;
+  padding: 19px 19px 0;
+  min-height: 70px;
 }
