+++ packages/client/src/pages/TopicsList/styles.module.scss
@@ -1,0 +1,37 @@
+@import '../../assets/styles/mixins';
+@import '../../assets/styles/variables';
+
+
+.container {
+  @include scrollbars;
+
+  display: flex;
+  flex-direction: column;
+  align-items: flex-end;
+  width: 100%;
+  height: 82vh;
+  overflow-y: auto;
+}
+
+.topicList {
+  display: flex;
+  flex-direction: column;
+  gap: 10px;
+  width: 100%;
+}
+
+.buttonContainer {
+  display: flex;
+  justify-content: flex-end;
+  width: 100%;
+}
+
+.button {
+  @include typography-xl;
+
+  border: none;
+  background: none;
+  padding: 5px;
+  letter-spacing: 2.4px;
+  color: $color-palette-green;
+}
